+++ resources/views/layouts/navigation/user-settings.blade.php
@@ -7,6 +7,9 @@
             <x-responsive-nav-link :href="route('settings')" iconClasses="fa-solid fa-screwdriver-wrench">
                 {{ __('Settings') }}
             </x-responsive-nav-link>
+            <x-responsive-nav-link :href="route('uploadedFiles')" iconClasses="fa-solid fa-folder-tree">
+                {{ __('Uploaded Files') }}
+            </x-responsive-nav-link>
         @endif
         <x-responsive-nav-link :href="route('profile.edit')" iconClasses="fa-solid fa-user-gear">
             {{ __('Profile') }}